refactor language module
