+++ src/main/scala/ai/privado/exporter/CollectionExporter.scala
@@ -37,7 +37,7 @@ import io.shiftleft.semanticcpg.language.*
 import org.slf4j.LoggerFactory
 import overflowdb.traversal.Traversal
 import io.shiftleft.semanticcpg.language.*
-import ai.privado.semantic.Language.*
+import ai.privado.semantic.language.*
 
 import scala.collection.mutable
 import scala.collection.mutable.ListBuffer
