+++ src/main/scala/ai/privado/exporter/ExporterUtility.scala
@@ -62,7 +62,7 @@ import ai.privado.model.exporter.CollectionEncoderDecoder.*
 import ai.privado.model.exporter.AndroidPermissionsEncoderDecoder.*
 import ai.privado.model.exporter.SinkEncoderDecoder.*
 import ai.privado.model.exporter.PropertyNodesEncoderDecoder.*
-import ai.privado.semantic.Language.{NodeStarterForSqlQueryNode, finder}
+import ai.privado.semantic.language.*
 import io.shiftleft.codepropertygraph.generated.{Cpg, Languages, NodeTypes}
 import ai.privado.utility.Utilities
 import ai.privado.utility.Utilities.{dump, getTruncatedText}
