+++ src/main/scala/ai/privado/exporter/HttpConnectionMetadataExporter.scala
@@ -24,7 +24,7 @@
 package ai.privado.exporter
 
 import ai.privado.cache.{AppCache, RuleCache}
-import ai.privado.semantic.Language.*
+import ai.privado.semantic.language.*
 import ai.privado.languageEngine.java.language.*
 import io.shiftleft.codepropertygraph.generated.Cpg
 import org.slf4j.LoggerFactory
