+++ src/test/scala/ai/privado/entrypoint/DynamicRuleMergerTest.scala
@@ -4,7 +4,7 @@ import ai.privado.cache.RuleCache
 import ai.privado.model.{CatLevelOne, ConfigAndRules, Constants, FilterProperty, Language, NodeType, RuleInfo}
 import ai.privado.testfixtures.JavaFrontendTestSuite
 import io.shiftleft.semanticcpg.language.*
-import ai.privado.semantic.Language.*
+import ai.privado.semantic.language.Language.*
 import ai.privado.languageEngine.java.language.module.NodeStarters
 
 class DynamicRuleMergerTest extends JavaFrontendTestSuite, DynamicRuleMerger {
